update predict function & label collection
